Flytt ut schema
